src/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-bigquery==3.17.0
google-cloud-vision==3.7.0
pytest==7.4.0
pytest-cov==3.0